// Konfigurator-Frontend/src/swagger/model/getGameMapDto.ts
/**
 * Konfigurator-aspnetcore-Backend
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MapSizeDto } from './mapSizeDto';
import { VectorDto } from './vectorDto';

export interface GetGameMapDto { 
    name?: string;
    username?: string;
    size?: MapSizeDto;
    tile?: string;
    barrier?: string;
    character?: string;
    target?: string;
    obstacles?: Array<VectorDto>;
    startPos?: VectorDto;
    endPos?: VectorDto;
    id?: number;
}